Code refactorings.
