+++ app/src/main/java/com/balicodes/quicksms/ExportImportActivity.kt
@@ -31,6 +31,7 @@ import android.support.v7.app.AppCompatActivity
 import android.util.Log
 import android.widget.Toast
 import com.balicodes.quicksms.entity.MessageEntity
+import com.balicodes.quicksms.model.SMSItem
 import com.balicodes.quicksms.viewmodel.MessageViewModel
 import kotlinx.android.synthetic.main.activity_export_import.*
 import org.jetbrains.anko.doAsync
