+++ app/src/main/java/com/balicodes/quicksms/SMSStatusBroadcastReceiver.kt
@@ -21,6 +21,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.util.Log
+import com.balicodes.quicksms.model.Recipient
 
 class SMSStatusBroadcastReceiver : BroadcastReceiver() {
