+++ app/src/main/java/com/balicodes/quicksms/viewmodel/MessageViewModel.kt
@@ -22,7 +22,7 @@ import android.arch.lifecycle.AndroidViewModel
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.Transformations
-import com.balicodes.quicksms.SMSItem
+import com.balicodes.quicksms.model.SMSItem
 import com.balicodes.quicksms.entity.MessageEntity
 import com.balicodes.quicksms.repository.MessageRepository
